+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## v1.0.1
+## v1.0.2
 
 <!--Releasenotes start-->
-- The default `RevealReactionVoiceLines` have been updated to include more varied voicelines by Melinoë.
+- Fixed potential issues with modded and vanilla cosmetics being equipped at the same time if the mod was temporarily disabled and the player returns from a Chaos Trial.
 <!--Releasenotes end-->
+
+## v1.0.1
+
+- The default `RevealReactionVoiceLines` have been updated to include more varied voicelines by Melinoë.
 
 ## v1.0.0